구조 재정의
