+++ pages/new/common.tsx
@@ -1,0 +1,10 @@
+export interface ItemProps {
+  id: string;
+  text_content: string;
+  url: string;
+  domain: string;
+  tags: string[];
+  timestamp: string;
+  source: string;
+  path: string;
+}
